Refactored project with decoupling
